+++ src/types.rs
@@ -1,6 +1,6 @@
 //! Data types used by the protocol.
 
 pub use oinq::request::{
-    Config, CrusherConfig, HogConfig, HostNetworkGroup, PigletConfig, ReconvergeConfig,
-    ResourceUsage,
+    Config, CrusherConfig, HogConfig, HostNetworkGroup, PigletConfig, Process, ReconvergeConfig,
+    ResourceUsage, TrafficFilterRule,
 };