+++ CHANGELOG.md
@@ -5,11 +5,21 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Types used in the protocol:
+  - `Process`
+  - `TrafficFilterRule`
+- `client::send_ok` and `client::send_err` methods to facilitate sending
+  responses from the client to the server.
+
 ## [0.1.1] - 2024-03-21
 
 ### Added
 
-- Types for agent configuration under `types`:
+- Types used in the protocol:
   - `Config`
   - `CrusherConfig`
   - `HogConfig`
@@ -34,5 +44,6 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - `client::handshake` implements the application-level handshake process for the
   client after a QUIC connection is established.
 
+[Unreleased]: https://github.com/petabi/review-protocol/compare/0.1.1...main
 [0.1.1]: https://github.com/petabi/review-protocol/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/petabi/review-protocol/tree/0.1.0